+++ digitalmall/digitalmall-portal/src/main/java/com/thezjh/digitalmallportal/dao/AddressMapper.java
@@ -16,5 +16,5 @@ public interface AddressMapper extends BaseMapper<UserAddress> {
      * @param userId
      * @return
      */
-    UserAddress getDefaultAddress(Long userId);
+    UserAddress getDefaultAddress(Integer userId);
 }
