+++ digitalmall/digitalmall-portal/src/main/java/com/thezjh/digitalmallportal/service/AddressService.java
@@ -13,5 +13,5 @@ public interface AddressService {
      * @param userId
      * @return
      */
-    UserAddress getDefaultAddress(Long userId);
+    UserAddress getDefaultAddress(Integer userId);
 }
